rename testdata files
