+++ src/MasterData/MasterData.BusinessLogic.Tests.Integration/SourceFormat/UpdateAsync_Validation_Should.cs
@@ -6,7 +6,7 @@ namespace DigitalLibrary.MasterData.BusinessLogic.Tests.Integration.SourceFormat
 
     using DigitalLibrary.MasterData.BusinessLogic.Implementations.SourceFormat;
     using DigitalLibrary.MasterData.DomainModel;
-    using DigitalLibrary.MasterData.Tests.TestData;
+    using DigitalLibrary.MasterData.Tests.TestData.SourceFormat;
 
     using FluentAssertions;
 
@@ -35,7 +35,7 @@ namespace DigitalLibrary.MasterData.BusinessLogic.Tests.Integration.SourceFormat
         }
 
         [Theory]
-        [ClassData(typeof(SourceFormat_UpdateAsync_Validation_TestData))]
+        [ClassData(typeof(UpdateAsync_Validation_TestData))]
         public async Task Throw_WhenInputIsInvalid(
             long id,
             string name,
