+++ src/MasterData/MasterData.Tests.TestData/SourceFormat/UpdateAsync_TestData.cs
@@ -1,11 +1,11 @@
-namespace DigitalLibrary.MasterData.Tests.TestData
+namespace DigitalLibrary.MasterData.Tests.TestData.SourceFormat
 {
     using System.Collections;
     using System.Collections.Generic;
     using System.Diagnostics.CodeAnalysis;
 
     [ExcludeFromCodeCoverage]
-    public class SourceFormat_UpdateAsync_TestData : IEnumerable<object[]>
+    public class UpdateAsync_TestData : IEnumerable<object[]>
     {
         IEnumerator IEnumerable.GetEnumerator()
         {
